+++ src/osi_cgraph.h
@@ -8,7 +8,7 @@ extern "C"
 
 extern double maxRowTime;
 extern int maxRowTimeIdx;
-extern int activePairwise, activeReadyToUse, inactivePairwise, inactiveReadyToUse, mixedPairwise;
+extern unsigned long int activePairwise, activeReadyToUse, inactivePairwise, inactiveReadyToUse, mixedPairwise;
 extern int success;
 
 /* lp should be a pointer to a OsiSolverInterface object */
